requirements.txt
# Verified against PyPI on 2026-04-15
# Pin these exactly to avoid surprise breakage

# UI — community-maintained since May 2025, pin firmly
chainlit==2.9.6

# LangChain ecosystem — verified latest as of Apr 2026
# NOTE: langchain 1.x removed ConversationalRetrievalChain.
#       We use the modern LCEL approach: create_history_aware_retriever + create_retrieval_chain
langchain==1.2.15
langchain-core==1.2.28
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-groq==1.1.2
langchain-huggingface==1.2.0

# Chroma — verified latest as of Apr 2026
# NOTE: Use langchain-chroma (dedicated package), not the community Chroma wrapper
langchain-chroma==0.2.4
chromadb==1.5.7

# PDF parsing — local, no API key needed
# unstructured[pdf] pulls in pdfminer, pillow, etc. automatically
unstructured[pdf]==0.17.2

# Embeddings (local, runs on your RTX 3050)
sentence-transformers==3.4.1

# Utilities
python-dotenv==1.1.0
